Add files via upload
Project Maxine and I did (with some part stock-source from the prof) as part of the last assignment before Christmas. 
Still a couple open questions of how certain stuff is supposed to work but yeah. 
requires base.h because of xcalloc-use (that way it also includes stdio.h and some more.)